refactor: 서비스 레이어 통합 테스트 패키지 분리
