+++ backend/src/test/java/com/festago/application/integration/ApplicationIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.festago.application;
+package com.festago.application.integration;
 
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
@@ -6,6 +6,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest(webEnvironment = WebEnvironment.NONE)
 @Transactional
-public abstract class ApplicationTest {
+public abstract class ApplicationIntegrationTest {
 
 }
